Fix: deduplicate parameter
