+++ qcodes/instrument_drivers/AlazarTech/acqusition_parameters.py
@@ -1,0 +1,66 @@
+from qcodes import Parameter
+
+class AcqVariablesParam(Parameter):
+    """
+    Parameter of an AcquisitionController which has a _check_and_update_instr
+    function used for validation and to update instrument attributes and a
+    _get_default function which it uses to set the AcqVariablesParam to an
+    instrument calculated default.
+
+    Args:
+        name: name for this parameter
+        instrument: acquisition controller instrument this parameter belongs to
+        check_and_update_fn: instrument function to be used for value
+            validation and updating instrument values
+        default_fn (optional): instrument function to be used to calculate
+            a default value to set parameter to
+        initial_value (optional): initial value for parameter
+    """
+
+    def __init__(self, name, instrument, check_and_update_fn,
+                 default_fn=None, initial_value=None):
+        super().__init__(name)
+        self._instrument = instrument
+        self._save_val(initial_value)
+        setattr(self, '_check_and_update_instr', check_and_update_fn)
+        if default_fn is not None:
+            setattr(self, '_get_default', default_fn)
+
+    def set(self, value):
+        """
+        Function which checks value using validation function and then sets
+        the Parameter value to this value.
+
+        Args:
+            value: value to set the parameter to
+        """
+        self._check_and_update_instr(value, param_name=self.name)
+        self._save_val(value)
+
+    def get(self):
+        return self._latest()['value']
+
+    def to_default(self):
+        """
+        Function which executes the default_fn specified to calculate the
+        default value based on instrument values and then calls the set
+        function with this value
+        """
+        try:
+            default = self._get_default()
+        except AttributeError as e:
+            raise AttributeError('no default function for {} Parameter '
+                                 '{}'.format(self.name, e))
+        self.set(default)
+
+    def check(self):
+        """
+        Function which checks the current Parameter value using the specified
+        check_and_update_fn which can also serve to update instrument values.
+
+        Return:
+            True (if no errors raised when check_and_update_fn executed)
+        """
+        val = self._latest()['value']
+        self._check_and_update_instr(val, param_name=self.name)
+        return True
